Refactored code to validate updatedCandidate.
