+++ src/lib/api/resume.ts
@@ -1,8 +1,5 @@
 // Dependencies.
-import type { NextRequest } from "next/server"
 import { NextResponse } from "next/server"
-import { validate } from "uuid"
-import type { z } from "zod"
 import {
 	type Accomplishment,
 	type Candidate,
@@ -12,92 +9,6 @@ import {
 	type Skill,
 	type SkillSet,
 } from "@/data/resume"
-
-// Validate UUID format.
-export function validateUuidFormat(uuid: string) {
-	if (!validate(uuid)) {
-		return NextResponse.json(
-			{
-				error: `The UUID ${uuid} isn’t valid.`,
-			},
-			{ status: 400 },
-		)
-	}
-
-	return null
-}
-
-// Validate data found by candidate ID.
-export function validateDataFoundByCandidateId<T>(
-	candidateId: string,
-	data: T | null,
-	dataName: string,
-) {
-	if (!data) {
-		return NextResponse.json(
-			{
-				error: `Couldn’t find the ${dataName} by candidateId (${candidateId}).`,
-			},
-			{ status: 404 },
-		)
-	}
-
-	return null
-}
-
-// Validate the request body is valid JSON.
-export async function parseRequestBody(request: NextRequest) {
-	let requestBody: unknown
-
-	try {
-		requestBody = await request.json()
-	} catch {
-		return NextResponse.json(
-			{ error: "The request body isn't valid JSON." },
-			{ status: 400 },
-		)
-	}
-
-	return requestBody
-}
-
-// Validate the request body against the schema.
-export function validateRequestBodyAgainstSchema<T>(
-	requestBody: unknown,
-	schema: z.ZodType<T>,
-) {
-	const parsedRequestBody = schema.safeParse(requestBody)
-
-	if (!parsedRequestBody.success) {
-		const issues = parsedRequestBody.error.issues.map((issue) => {
-			const formatted: { message: string; path?: string } = {
-				message: issue.message,
-			}
-
-			if (issue.code === "unrecognized_keys") {
-				formatted.message = `The schema doesn't allow these fields: ${issue.keys.join(", ")}.`
-				return formatted
-			}
-
-			if (issue.path.length > 0) {
-				formatted.path = issue.path.join(".")
-			}
-
-			return formatted
-		})
-
-		return NextResponse.json(
-			{ error: "The request body doesn’t match the schema.", issues },
-			{ status: 400 },
-		)
-	}
-
-	return parsedRequestBody.data
-}
-
-//
-// **
-//
 
 // Validate the candidate ID.
 export function validateCandidateId(candidateId: string): NextResponse | null {
